pyproject.toml
[project]
name = "easyobs"
version = "0.1.0"
description = "EasyObs: lightweight observability console for LLM agents (OTLP/HTTP, multi-tenant, RBAC)"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
classifiers = [
  "License :: Other/Proprietary License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "aiosqlite>=0.20.0",
  "httpx>=0.27.0",
  "opentelemetry-proto>=1.28.0",
  "protobuf>=5.0.0",
  "argon2-cffi>=23.1.0",
  "pyjwt>=2.9.0",
  # Used by `easyobs reset-data --force` to enumerate / kill leftover
  # uvicorn workers (especially Windows --reload orphans whose parent has
  # exited but child still holds the LISTEN socket + catalog handle).
  "psutil>=5.9.0",
  # MIT-licensed xlsx reader/writer used by the Golden Set upload service
  # (12 §9). Picked over pandas/SheetJS to keep the dependency surface
  # small and the licensing unambiguous (see 12 §9.1).
  "openpyxl>=3.1.0",
  # FastAPI ``UploadFile`` requires multipart support to land in memory.
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
agent = [
  "opentelemetry-api>=1.28.0",
  "opentelemetry-sdk>=1.28.0",
  "opentelemetry-exporter-otlp-proto-http>=1.28.0",
]
langchain = [
  "opentelemetry-api>=1.28.0",
  "opentelemetry-sdk>=1.28.0",
  "opentelemetry-exporter-otlp-proto-http>=1.28.0",
  "langchain-core>=0.3.0",
]
pricing = ["tokencost>=0.1.13"]
judges = [
  "openai>=1.12.0",
  "anthropic>=0.39.0",
  "google-genai>=1.0.0",
  "google-cloud-aiplatform>=1.60.0",
  "boto3>=1.34.0",
]
dev = ["pytest>=8.3.0", "pytest-asyncio>=0.24.0", "ruff>=0.8.0"]

[project.scripts]
easyobs = "easyobs.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/easyobs", "src/easyobs_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
